Fix the homebrew formula for govc and bump the version to 0.19
The upstream project changed how the packaging is done. It's now a
single gzipped binary instead of a gzipped tarball.

In order to accomodate this change, the formula needs to move the
gunzipped binary from `govc_darwin_amd64` to `govc`.

While here, bump the govc version from 0.18 to 0.19.

This change fixes Issue #2.